started list design
i think there were some itty bitty touches to queue_design.md, but yeah.

list_design.md begun for realsies. i'm not super sure of this design, and i think that's because of the efficiency thing, i feel silly and not even sure how big the efficiency concern is. it's unfamiliar. i believe the design still works.

may also be concerned about the simplicity of the design. and the queue design. going back and making the queue a linked queue is a much simpler answer, but i had fun coming up what i came up with. and even then i have 3 other versions of it.

ahahgahghahgaghgah