complexity analysis
???? ?

not sure about this one. main concern is line 157: would calling another function have an effect on space? even if that isn't dependent on the input. i'm going solely off the geeksforgeeks article at this point in time aaagh.

everything is O(1), that concerns me. i'm not sure i 100% understand this mathematically and i'm not feeling like i have the best references atm. that's ok.