began advanced hash
simple hash [Hashtable until now] has been moved to files simplehash.cpp and .h. new hash is in files hashtable.cpp and .h. kinda confusing as a point of transition, but should make sense in the future hopefully. this is the Real Deal, and that one is the simpler version. right? right.