removed deleteVertex()
realized i don't actually need this, nor does it completely make sense. it suppose it could be used to free up memory space, but for the game itself, it's better to remove the arcs leading to the node instead.

also you kind of Have to remove the arcs leading to the node. i don't have a simple way of doing this with how i organize arcs, there'd be a bunch of loose ends.