removed list references and fixed stuff
so i had a scare and thought that i needed to use a list instead of a vector. i learned a couple things, but ultimately that i am a fool. my deletePath() method or whatever deleted an entire node, not just an edge. fixed that, now all the list implementations are reverted, and we are good to go.