added objectives, updated error messages, BEGAN WRITING!!!
i actually figured out what i want the wizard tone to be like. anyways.

player now has objectives, which can be set in a callback or literally wherever.
- private objectives variable
- set objective method
- read objective method
- check objective method: returns bool for completion.

game loop altered to include checking on objectives.

wizardville.cpp actually has a bit of a game in it so far! a nonlinear traversable map with updating passages and text.